Add blank line after rake task output
The lokaise rake tasks call print rather than puts.  So a blank line
after is needed for Heroku build output to look as expected.  Withtout
the blank line the `----->` expected to denote the execution of a new
task doesn't occur on a new line and results in easy to miss meaning in
the build output.